decode buffers as 'latin1' by default
This preserves "binary" data, and allows for treating text as
Windows-1252, even if declared as ISO-8859-1.

For now, don't do the actual patching at this layer, though; that would
corrupt true binary data. Export a function that does it, which can be
called after separating multi-part content.

Also, both tab and space are allowed for header folding.